Code refactored in a major way
The main refactoring was to change how addTask and RemoveTask Function on a fundamental level to support more complex GUIs, The controller class has also been changed significantly. The original GUI has been scrapped as it is no longer functionable due to how major this refactoring was. To create a new task type, one must create a corresponding factory, and Modify the switch statement withing the factoryCreator (this felt unavoidable due to us using enumerations.)